requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0
tqdm>=4.65.0
pillow>=9.5.0
numpy>=1.24.0
